refactor(comments): share reply page handlers
